Update README.md
The url provided was not working, instead https://sentry.io is working so changed. (If a wrong url is provided then it not upload to sentry).
Now one important step in this doc which was missing it adding sourcemap true in build options of vite config. So, i've added it. 
We should add it as dev dependancy when installing through yarn package manager.