peerDependencies: allow more vite versions
I had to do this change to use `vite-plugin-sentry` on a project using `vite@^2.7.1` — well, had to add the `|| ^2.7.1` part for that. Another project was using `vite@^2.9.0` and that's why I added the `|| ^2.9.0`.  Both cases, `^2.7.1 || ^2.9.0`, have been tested.

(It could make sense to add a `|| ^2.8.0` as well, because then every major version between 2.6 and 3.0 would be allowed. But I just haven't tested that case so I'm not including that change.)